fix script

+++ preferences.sh
@@ -12,9 +12,19 @@ printf "\t-> load guake on startup \n"
 	ln -s /usr/share/applications/guake.desktop /etc/xdg/autostart/
 
 printf "\t-> copying awesome config \n"
+	if [ ! -d /home/$1/.config/awesome ]; then
+		mkdir /home/$1/.config/awesome
+	fi
 	cp -r dl/awesome-copycats/* /home/$1/.config/awesome
 	cp pref/rc.lua /home/$1/.config/awesome	
 
 printf "\t-> setting awesome config \n"
-	rm /etc/xdg/awesome/rc.lua
-	ln -s /home/$1/.config/awesome/rc.lua rc.lua 
+	if [ ! -f /etc/xdg/awesome/rc.lua]; then
+		rm /etc/xdg/awesome/rc.lua
+	fi
+	ln -s rc.lua /home/$1/.config/awesome/rc.lua
+	echo "guake&
+
+source /home/$1/.zshrc
+
+exec awesome" > /home/$1/.xinitrc
